backend/requirements.txt
asgiref==3.7.2
dj-config-url==0.1.1
Django==3.2.23
django-cors-headers==4.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
